# src/docs/swagger.yaml
definitions:
  dto.ReqCreateUpdateEntry:
    properties:
      description:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      tag_list:
        items:
          type: integer
        type: array
      time_end:
        type: string
      time_start:
        type: string
    required:
    - time_end
    - time_start
    type: object
...
  dto.ReqUserSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.ReqUserSignUp:
    properties:
      about:
        type: string
      avatar_id:
        type: integer
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.RespEntry:
    properties:
      description:
        type: string
      duration:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      tag_list:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      time_end:
        type: string
      time_start:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.RespUser:
    properties:
      about:
        type: string
      avatar_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  echo.HTTPError:
    properties:
      message: {}
    type: object
  models.Tag:
    properties:
      about:
        type: string
      color:
        type: string
      id:
        type: integer
      name:
        type: string
      userID:
        type: integer
    type: object
  models.User:
    properties:
      about:
        type: string
      avatar_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
  pkg.Response:
    properties:
      body: {}
    type: object
info:
  contact: {}
paths:
  /auth:
    get:
      description: check user auth
      produces:
      - application/json
      responses:
        "200":
          description: success auth
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Auth
      tags:
      - auth
  /entry/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an entry
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find entry with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Delete an entry
      tags:
      - entry
    get:
      consumes:
      - application/json
      description: Get entry by id
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get entry
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespEntry'
              type: object
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Show a post
      tags:
      - entry
  /entry/create:
    post:
      consumes:
      - application/json
      description: Create entry
      parameters:
      - description: entry info
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateEntry'
      produces:
      - application/json
      responses:
        "200":
          description: success update entry
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespEntry'
              type: object
        "400":
          description: bad req
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Create entry
      tags:
      - entry
  /entry/edit:
    post:
      consumes:
      - application/json
      description: Update an entry
      parameters:
      - description: entry info
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateEntry'
      produces:
      - application/json
      responses:
        "200":
          description: success update entry
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespEntry'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Update an entry
      tags:
      - entry
  /entry/usr:
    get:
      description: Get user entries or get user entries for a day
      parameters:
      - description: day for events
        in: query
        name: day
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success get entries
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/dto.RespEntry'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get user entries
      tags:
      - entry
...
  /logout:
    post:
      description: user logout
      produces:
      - application/json
      responses:
        "204":
          description: success logout, body is empty
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Logout
      tags:
      - auth
...
  /signin:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: success sign in
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespUser'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: invalid password
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SignIn
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: user sign up
      parameters:
      - description: user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.ReqUserSignUp'
      produces:
      - application/json
      responses:
        "201":
          description: user created
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespUser'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: email already exists
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SignUp
      tags:
      - auth
...
swagger: "2.0"